Create README.md
Pet Inventory System which will help visualize sales data pertaining to a particular store such sales per month order price, supply amount and employee payout. The system also gets the details of medical record and immunization record of pet which will be useful for owner to produce it to customer. Apart from that system also maintains sensitive data of the user for security purpose using encryption techniques. The system keep track of all the products and order line details using the audit table which help the user to keep track of stock of the products as and when required. The system has the capability to generate the sales reports. It will help the shop owner to make decisions based on the historical data and manage the business efficiently.